pyproject.toml
[project]
name = "lyra-api-v3"
version = "0.1.0"
description = "掩码生成模型训练系统 - 基于ADE20K数据集的语义分割"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "tensorboard>=2.13.0",
    "scikit-learn>=1.3.0",
    "albumentations>=1.3.0",
    "segmentation-models-pytorch>=0.3.2",
]
